requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.8
narwhals==2.18.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
plotly==6.6.0
psycopg2==2.9.11
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.48
threadpoolctl==3.6.0
typing_extensions==4.15.0
